personal_portfolio
A portfolio website is a personal website that showcases the skills, work, and achievements of an individual, typically a professional or a creative, in a visually appealing and organized manner. It serves as a digital portfolio or online resume, allowing individuals to present their work and capabilities to potential employers, clients, or collaborators.